Refactor signin/signout buttons and rename components
